requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2023.1
Django==4.2.4
django-crispy-forms==2.1
Pillow==10.1.0
sqlparse==0.4.4
tzdata==2023.3
